requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
